refactor: Move LaTeX and Typst integration into package extensions
Typstry and tectonic_jll are now [weakdeps] — MakieTeX no longer pulls
them in by default. Users opt in by loading the respective package:

  using MakieTeX, tectonic_jll  # activates MakieTeXLaTeXExt
  using MakieTeX, Typstry       # activates MakieTeXTypstExt

The LaTeX/Typst handler struct definitions (LaTeX, FullLaTeX, Typst,
FullTypst) stay in core so they exist regardless of which extension is
loaded. Concrete compile_text dispatches and the engine-running code
live in the extensions. The shared place_text! / bbox / crop helpers
move to src/pdf_text_handler.jl.

Legacy TypstDocument / CachedTypst struct definitions stay in core; the
constructors that need Typst compilation move to MakieTeXTypstExt.

LaTeXStrings remains a hard dep (lightweight, transitively required by
Typstry too).
