+++ src/MakieTeX.jl
@@ -2,7 +2,7 @@ module MakieTeX
 
 using Makie
 
-using Colors, LaTeXStrings, Typstry
+using Colors, LaTeXStrings
 using Base64
 
 # Patch for Makie.jl `@Block` macro error
@@ -13,14 +13,11 @@ using Makie.GeometryBasics: origin, widths
 using Makie.Observables
 using DocStringExtensions
 
-using Poppler_jll, Ghostscript_jll, Glib_jll, tectonic_jll
+using Poppler_jll, Ghostscript_jll, Glib_jll
 using Rsvg, Cairo
 
-# Define some constants for configuration
-"The current `TeX` engine which MakieTeX uses."
-const CURRENT_TEX_ENGINE = Ref{Cmd}(`lualatex`)
-"Default margins for `pdfcrop`.  Private, try not to touch!"
-const _PDFCROP_DEFAULT_MARGINS = Ref{Vector{UInt8}}([0,0,0,0])
+# Default margins for `pdfcrop`.  Private, try not to touch!
+const _PDFCROP_DEFAULT_MARGINS = Ref{Vector{UInt8}}([0, 0, 0, 0])
 "Default density when rendering images"
 const RENDER_DENSITY = Ref(3)
 
@@ -28,12 +25,12 @@ const RENDER_DENSITY = Ref(3)
 include("types.jl")
 
 include("rendering/pdf_utils.jl")
-include("rendering/typst.jl")
 include("rendering/pdf.jl")
 include("rendering/svg.jl")
 
-include("text_integration.jl")
-include("typst_integration.jl")
+include("pdf_text_handler.jl")
+include("latex_handler.jl")
+include("typst_handler.jl")
 
 export Cached
 export TypstDocument, CachedTypst
@@ -44,63 +41,5 @@ export LaTeX, FullLaTeX
 export Typst, FullTypst
 
 export LaTeXStrings, LaTeXString, latexstring, @L_str
-export Typstry, TypstString, @typst_str
 
-"Try to write to `engine` and see what happens"
-function try_tex_engine(engine::Cmd)
-    try
-        fd = open(engine; write = true)
-        write(fd, "\n")
-        close(fd)
-        return nothing
-    catch err
-        println("The TeX engine $(CURRENT_TEX_ENGINE[]) failed.")
-        return err
-    end
-end
-
-"Checks whether the default latex engine is correct"
-function __init__()
-
-    # First, determine latex engine support
-    latexmk = Sys.which("latexmk")
-    if isnothing(latexmk)
-        @warn """
-        MakieTeX could not find `latexmk` on your system!
-        If you want to use the `luatex` engine, or any local or non-standard
-        packages, then please install `latexmk` and ensure that it is on `PATH`.
-
-        Defaulting to the bundled `tectonic` renderer for now.
-        """
-        CURRENT_TEX_ENGINE[] = `tectonic`
-    else
-        t1 = try_tex_engine(CURRENT_TEX_ENGINE[]) # by default `lualatex`
-
-        if !isnothing(t1)
-
-            @warn("""
-                The specified TeX engine $(CURRENT_TEX_ENGINE[]) is not available.
-                Trying pdflatex:
-                """
-            )
-    
-            CURRENT_TEX_ENGINE[] = `pdflatex`
-        else
-            return
-        end
-    
-        t2 = try_tex_engine(CURRENT_TEX_ENGINE[])
-        if !isnothing(t2)
-    
-            @warn "Could not find a TeX engine; defaulting to bundled `tectonic`"
-            CURRENT_TEX_ENGINE[] = `tectonic`
-        else
-            return
-        end
-    
-    end
-
-    return
-end
-
-end # document
+end # module
